feat: add suport for SSR projects
This commit updates scripts and configs to support SSR style projects.

Changes to be committed:
renamed:    commitlint.config.js -> commitlint.config.cjs
modified:   package-lock.json
modified:   package.json
modified:   packageScripts/postinstall.mjs
renamed:    scripts/generateDocs.js -> scripts/generateDocs.mjs
deleted:    scripts/postCss.js
new file:   scripts/postCss.mjs
modified:   scripts/pre-commit.mjs
