+++ package.json
@@ -13,6 +13,7 @@
     "type": "git",
     "url": "https://github.com/AlaskaAirlines/auro-flight"
   },
+  "type": "module",
   "main": "index.js",
   "license": "Apache-2.0",
   "engines": {
@@ -33,6 +34,7 @@
     "@alaskaairux/design-tokens": "^3.15.5",
     "@alaskaairux/icons": "^4.30.1",
     "@aurodesignsystem/design-tokens": "^4.1.1",
+    "@aurodesignsystem/auro-library": "^2.2.5",
     "@aurodesignsystem/eslint-config": "^1.3.0",
     "@aurodesignsystem/webcorestylesheets": "^5.0.5",
     "@commitlint/cli": "^18.4.3",
@@ -137,7 +139,7 @@
     "build:ci": "npm-run-all sweep build",
     "build:api": "wca analyze 'src/auro-flight*.js' --outFiles docs/api.md",
     "build:dev:assets": "npm-run-all build:sass:component postCss:component sass:render",
-    "build:docs": "node scripts/generateDocs.js",
+    "build:docs": "node scripts/generateDocs.mjs",
     "build:sass": "npm-run-all build:sass:component postCss:component sass:render",
     "build:sass:component": "sass --no-source-map src:src",
     "build:watch": "nodemon -e scss,js --watch src --exec npm run build:dev:assets",
@@ -146,10 +148,10 @@
     "scssLint": "stylelint \"./src/**/*.scss\"",
     "dev": "concurrently --kill-others 'npm run build:watch' 'npm run serve'",
     "dist:js": "copyfiles -u 1 -V './src/**/*.js' ./dist",
-    "esLint": "./node_modules/.bin/eslint src/**/*.js",
+    "esLint": "eslint src/auro-*.js",
     "linters": "npm-run-all scssLint esLint",
     "preCommit": "node scripts/pre-commit.mjs",
-    "postCss:component": "node ./scripts/postCss.js",
+    "postCss:component": "node scripts/postCss.mjs",
     "postinstall": "node packageScripts/postinstall.mjs",
     "sass:render": "sass-render src/*.css -t ./scripts/staticStyles-template.js",
     "serve": "web-dev-server --open demo/ --node-resolve --watch",
