+++ scripts/generateDocs.mjs
@@ -1,7 +1,14 @@
-const path = require('path');
-const markdownMagic = require('markdown-magic');
-const fs = require('fs');
-const https = require('https');
+/**
+ * This version is sans dependencies
+ * Create for use with the SSR updates
+ */
+
+import path from 'path';
+import markdownMagic from 'markdown-magic';
+import fs from 'fs';
+import https from 'https';
+
+const __dirname = new URL('.', import.meta.url).pathname;
 
 const readmeTemplateUrl = 'https://raw.githubusercontent.com/AlaskaAirlines/WC-Generator/master/componentDocs/README.md';
 const dirDocTemplates = './docTemplates';
@@ -11,29 +18,29 @@ const readmeFilePath = dirDocTemplates + '/README.md';
  * Extract NPM, NAMESPACE and NAME from package.json
  */
 
- function nameExtraction() {
-  const packageJson = fs.readFileSync('package.json', 'utf8', function(err, data) {
-    if (err) {
-      console.log('ERROR: Unable to read package.json file', err);
-    }
-  })
+function nameExtraction() {
+ const packageJson = fs.readFileSync('package.json', 'utf8', function(err, data) {
+   if (err) {
+     console.log('ERROR: Unable to read package.json file', err);
+   }
+ })
 
-  pName = JSON.parse(packageJson).name;
+ let pName = JSON.parse(packageJson).name;
 
-  let npmStart = pName.indexOf('@');
-  let namespaceStart = pName.indexOf('/');
-  let nameStart = pName.indexOf('-');
+ let npmStart = pName.indexOf('@');
+ let namespaceStart = pName.indexOf('/');
+ let nameStart = pName.indexOf('-');
 
-  let result = {
-    'npm': pName.substring(npmStart, namespaceStart),
-    'namespace': pName.substring(namespaceStart + 1, nameStart),
-    'namespaceCap': pName.substring(namespaceStart + 1)[0].toUpperCase() + pName.substring(namespaceStart + 2, nameStart),
-    'name': pName.substring(nameStart + 1),
-    'nameCap': pName.substring(nameStart + 1)[0].toUpperCase() + pName.substring(nameStart + 2)
-  };
+ let result = {
+   'npm': pName.substring(npmStart, namespaceStart),
+   'namespace': pName.substring(namespaceStart + 1, nameStart),
+   'namespaceCap': pName.substring(namespaceStart + 1)[0].toUpperCase() + pName.substring(namespaceStart + 2, nameStart),
+   'name': pName.substring(nameStart + 1),
+   'nameCap': pName.substring(nameStart + 1)[0].toUpperCase() + pName.substring(nameStart + 2)
+ };
 
-  return result;
-};
+ return result;
+}
 
 /**
  * Replace all instances of [npm], [name], [Name], [namespace] and [Namespace] accordingly
@@ -65,7 +72,7 @@ function formatTemplateFileContents(content, destination) {
    * Write the result to the destination file
    */
   fs.writeFileSync(destination, result, { encoding: 'utf8'});
-};
+}
 
 function formatApiTableContents(content, destination) {
   const nameExtractionData = nameExtraction();
@@ -80,31 +87,9 @@ function formatApiTableContents(content, destination) {
 
   fs.writeFileSync(destination, result, { encoding: 'utf8'});
 
-  fs.readFile('./demo/apiExamples.md', 'utf8', function(err, data) {
-    formatTemplateFileContents(data, './demo/apiExamples.md');
+  fs.readFile('./demo/api.md', 'utf8', function(err, data) {
+    formatTemplateFileContents(data, './demo/api.md');
   });
-}
-
-/**
- * If auroLabs project, include auroLabs documentation in `./README.md`
- */
-
-function processLabsReadmeContent() {
-  let nameExtractionData = nameExtraction();
-
-  if (nameExtractionData.npm === '@aurolabs') {
-    const callbackAurolabs = function(updatedContent, outputConfig) {
-      console.log('Readme updated to reference AuroLabs content.');
-    };
-
-    const configAurolabs = {
-      matchWord: 'AUROLABS-GENERATED-CONTENT'
-    };
-
-    const markdownPathAurolabs = path.join(__dirname, '../README.md');
-
-    markdownMagic(markdownPathAurolabs, configAurolabs, callbackAurolabs);
-  }
 }
 
 /**
@@ -113,7 +98,6 @@ function processLabsReadmeContent() {
 
 function processReadme() {
   const callback = function(updatedContent, outputConfig) {
-    processLabsReadmeContent()
 
     if (fs.existsSync('./README.md')) {
       fs.readFile('./README.md', 'utf8', function(err, data) {
@@ -132,8 +116,6 @@ function processReadme() {
   const markdownPath = path.join(__dirname, '../docTemplates/README.md');
 
   markdownMagic(markdownPath, config, callback);
-
-  processLabsReadmeContent();
 }
 
 /**
@@ -161,40 +143,18 @@ function processDemo() {
   markdownMagic(markdownPath, configDemo, callback);
 }
 
-function processDot() {
-
-  const callback = function(updatedContent, outputConfig) {
-    if (fs.existsSync('./demo/dotCompliance.md')) {
-      fs.readFile('./demo/dotCompliance.md', 'utf8', function(err, data) {
-        formatTemplateFileContents(data, './demo/dotCompliance.md');
-      });
-    } else {
-      console.log('ERROR: ./demo/dotCompliance.md file is missing');
-    }
-  };
-
-  const configDemo = {
-    matchWord: 'AURO-GENERATED-CONTENT',
-    outputDir: './demo'
-  };
-
-  const markdownPath = path.join(__dirname, '../docs/partials/dotCompliance.md');
-
-  markdownMagic(markdownPath, configDemo, callback);
-}
-
 /**
  * Compiles `./docTemplates/apiExamples.md` -> `./demo/apiExamples.md`
  */
 
 function processApiExamples() {
   const callback = function(updatedContent, outputConfig) {
-    if (fs.existsSync('./demo/apiExamples.md')) {
-      fs.readFile('./demo/apiExamples.md', 'utf8', function(err, data) {
-        formatApiTableContents(data, './demo/apiExamples.md');
+    if (fs.existsSync('./demo/api.md')) {
+      fs.readFile('./demo/api.md', 'utf8', function(err, data) {
+        formatApiTableContents(data, './demo/api.md');
       });
     } else {
-      console.log('ERROR: ./demo/apiExamples.md file is missing');
+      console.log('ERROR: ./demo/api.md file is missing');
     }
   };
 
@@ -203,7 +163,7 @@ function processApiExamples() {
     outputDir: './demo'
   };
 
-  const markdownPath = path.join(__dirname, '../docs/partials/apiExamples.md');
+  const markdownPath = path.join(__dirname, '../docs/partials/api.md');
 
   markdownMagic(markdownPath, config, callback);
 }
@@ -242,5 +202,5 @@ function copyReadmeLocally() {
  * Run all the actual document generation
  */
 copyReadmeLocally();
+processApiExamples();
 processDemo();
-processDot();
